Ceph for standard: Puppet functional changes
In order to enable Openstack's helm charts on StarlingX we need a
distributed persistent storage for Kubernetes that leverages our
existing storage configurations. For this stage we will enable CEPH's
RBD to work with Kubernetes RBD provisioner through a new Helm chart.

Since RBD will be the persistent storage solution, CEPH support has to
be extended to the 1, 2 node and Standard configurations.

This commit enables CEPH on a standard configuration without dedicated
storage by allowing users to add the 3rd Ceph monitor to a worker node.
It implements the puppet functional part that enables ceph to work on
a standard deployment.

Details:
* Update ceph.conf with the new monitor by removing unused ones
* Reserve space on any worker for ceph-mon-lv, this will make sure
  that user is able to configure a ceph-mon at any time without the
  need to extend platform-vg.
* In System Inventory make sure the correct IP addresses are
  passed to the puppet yaml's.

Change-Id: I7b8a01a7f2cf18f20e3509c2416f038580b4a071
Implements: containerization-2002844-CEPH-persistent-storage-backend-for-Kubernetes
Story: 2002844
Task: 28723
Depends-On: https://review.openstack.org/629166/
Signed-off-by: Ovidiu Poncea <Ovidiu.Poncea@windriver.com>