Ceph for standard: Semnatic checks
In order to enable Openstack's helm charts on StarlingX we need a
distributed persistent storage for Kubernetes that leverages our
existing storage configurations. For this stage we will enable CEPH's
RBD to work with Kubernetes RBD provisioner through a new Helm chart.

Since RBD will be the persistent storage solution, CEPH support has to
be extended to the 1, 2 node aand Standard configurations.

This commit adds multiple semantic checks to make sure that Ceph on
a standard deployment works as expected.

Change-Id: I31786fa78e8c926a57077bb0a25f1e4cbf543cab
Co-Authored-By: Stefan Dinescu <stefan.dinescu@windriver.com>
Implements: containerization-2002844-CEPH-persistent-storage-backend-for-Kubernetes
Story: 2002844
Task: 28723
Depends-On: https://review.openstack.org/629512
Signed-off-by: Ovidiu Poncea <Ovidiu.Poncea@windriver.com>