Add cluster-host IP for worker node in /etc/hosts
mtcClient monitors the cluster-host interface and looks up the
cluster-host IP address. Since the cluster-host network is not
served by dnsmasq, the cluster-host IP address is not available
in the dnsmasq.hosts file.
This commit add the cluster-host IP address to the /etc/hosts file
for all hosts. The host entry for the local host is added.
The other worker hosts cannot be retrieved from the hosts file.

Story: 2004273
Task: 28899
Change-Id: I80937e8809ed5d40cf57d97d34eaab6afd7d6400
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>