Support 'True' and 'False' input for system commands
add args choice 'True' and 'False' for https_enabled and sdn_enabled
in 'system modify' and for dynamic in 'system network-add' command
to support input 'True' and 'False'.

make https_enabled and sdn_enabled to lower case in sys-api controller

change choices to lower case when passing to sysinv-api and
sysinv-controller to make sure functional work.

Closes-bug: 1812268

Change-Id: If6411852edfc970ee1d749e4d4ccb8d045767bbc
Signed-off-by: sunausti <sunausti@starlingx.com>