Ceph for standard: Add deployment model to Systsm Inventory cluster API
In order to enable StarlingX Dashboard to display and allow configuration
of Ceph we meed a new field, deployment_model, added to the 'cluster' API
reponses. This field is generated at runtime based on the configured ceph
monitors and type of system.

Four options have been defined for deplyoment_model:
o storage-nodes: OSDs are depolyed on storage nodes. Main trigger for
  this model is the creation of a node with storage personality.
o controller-nodes: OSDs are deployed on controller nodes. Main trigger is
  creation of Ceph monitor on a worker node. This is the implicit model on
  AIO Duplex.
o aio-sx: OSDs are deployed on a simplex controller.
o undefined: before either a storage node is installed or a compute monitor
  is created on a worker.

This commit also enables storage profiles on controllers based on storage
model and adds the new field to CLI output of 'system cluster-*' commands.

Change-Id: I4d5cb6120abb176025857ae9749e8eb6ca121475
Implements: containerization-2002844-CEPH-persistent-storage-backend-for-Kubernetes
Story: 2002844
Task: 29113
Signed-off-by: Ovidiu Poncea <Ovidiu.Poncea@windriver.com>