Support ovs in container
stx support deploy with or without kubernetes. For the case kuberntes
enabled, most openstack components run in kuberntes containers. But ovs
runs on the host.

This patch is to support ovs running in container for kubernetes enabled
case, and make it the default. For kuberntes diabled case, ovs-dpdk is
the default.

Change-Id: I111865de17fc82a1e7c269fb97f53a181fb382b2
Story: 2004649
Task: 28610