Fix multi-net interface configuration
On the worker and storage nodes, the mgmt interface should be
set to DHCP and the cluster host alias should be set to use static IP.
However, the recent fix for this added DHCP to an alias interface which
is not supported.  This was causing the management network to not be
configured after the host was rebooted and the interfaces were
reconfigured by Puppet.

The interface configuration has been updated to ensure that DHCP
networks are configured against the parent device and all additional
static addresses will be configured as alias interfaces.

There is also a few additional issues that have been fixed:
  - multinetting with different address families
  - traffic controls applied to parent interface
  - address scope for aliases on AIO-SX with dedicated interfaces
  - setting of the gateway address on multi-net interfaces

Story: 2004273
Task: 27826

Change-Id: If8e354bf41cbc1c11966a4b43b8a058c33580c49
Signed-off-by: Matt Peters <matt.peters@windriver.com>