MTU set to default even if it is configured
The MTU value was set to the default value instead of the configured
value for the interface.
This commit sets the MTU value to the configured for the interface.

The mtu value is also used to calculate the read/write size for nfs
mounts. The commit also ensures that it is set to the correct value.

Closes-Bug: 1811404
Change-Id: Ieb3c6d989726075af205823dc1d570c00a24c28d
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>