Move gnocchi and ceilometer static configs
Move all gnocchi and ceilometer static configurations from the
overrides to the Armada manifest.

This is being done so we have a consistent way of managing
containerized openstack configurations. Static configurations will
be located in the Armada manifest and dynamic configuration will be
located in the overrides files.

Story: 2003909
Task: 29535

Change-Id: Ieab861cb1751146b70f722e70b8f89d81c0ed9a5
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>