Boost sm process priority in VBox environment
There is an instance that sm claimed its main thread ran sluggish
as some critical timer run behind the scheuled timing.
The issue could prevent the sm from scheduling services.
As the result, the controller could fail to enable.

The issue was found only on vbox labs on AIO-SX, the fix is to boost
sm process priority to nice value -10 from current -2.

Closes-Bug: 1816764
Depends-On: https://review.openstack.org/638664
Change-Id: Iafa17b1c47d65cc7394552ea1c8e7a78398e4869
Signed-off-by: Bin Qian <bin.qian@windriver.com>