Fix to bring up parent interface before its alias
The alias interface may at times be brought up before its parent
interface (ethernet or vlan) depending on how the interface config
files are listed in the filesystem.
The fix is to sort the interface config files such that the alias
are listed after the parent before bringing them down and up.

Partial-Bug: 1817593

Change-Id: I83b88e1587c6468d85e14a50de934908586cbc9b
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>