CEPH: Adding OSD storage devices no longer need nodes to be locked
Current procedur fore adding OSDs on AIO-DX and Standard w/a storage
nodes (2+2) requires controllers to be locked, OSDs added, node unlocked,
controllers swacted then lock/unlock is needed for the other controller.
This is too complex and not cut for production.

This commit simplify OSD management by enabling runtime addition on a
node that is 'available' thus avoiding the complex procedure described
above.

The runtime is done by running the ceph OSD puppet manifests when OSDs
are added. Ceph OSDs are stopped, manifests are executed, OSDs are
started and status is reported back to sysinv.

Monitoring is done in a new column called 'state' with 4 values:
1. 'configuring' - puppet manifests are running;
2. 'configuring-on-unlock - node is locked and manifests will be run
   on unlock;
3. 'configured' - once configuration is successful, #1&#2 above will
   turn into #3;
4. 'configuration-failed' - error state if manifests fail, this may be
   corrected by a lock/unlock of the node or a reinstall;

The 'state' column can be checked with 'system host-stor-list
<node_id_or_hostname>'.

Change-Id: Iccaaa0728a43936cd4c6e1ed590e01ba8f0fbaaa
Story: 2002844
Task: 29795
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>