Provide helm chart for containerized Barbican service
This commit introduces Barbican as an OpenStack Application.
Adding the Armada manifest for Barbican and updating the helm
plugin with correct values to bring up containerized Barbican.

Also fixing the typo in memcache_secret_key for all the services
in order to leverage authtoken cache encryption key feauture.

Change-Id: I41c85969fe0554399721abb6efa73148bdc84cd4
Story: 2003108
Task: 29969
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>