Update py35 jobs to py37
The upstream OpenStack default node type changed yesterday
from Ubuntu Xenial to Bionic. We have a number of jobs pegged
directly to py35 which is no longer available in Bionic.

We can either pin the py35 jobs to Xenial nodes or update them
to py36 or py37.  OpenStack is supporting both of those for the
Stein release, since StarlingX does not have any position on
Python 3 versions (yet!) I am suggesting we move straight to py37
and in the case of issues now make these jobs non-voting until
any incompatibilities are resolved.

Also:
* remove pike-era constraints files, we should use master now.
* force pysycopg2>=2.7.7
* add pycrypto to sysinv/sysinv/sysinv/requirements.txt

Change-Id: I4a86e211b2f4f323865cf6758302e6ac10b63890
Signed-off-by: Dean Troyer <dtroyer@gmail.com>