Adjust number of mariadb server pods
For the AIO-DX case (and the dual-controller-with-storage case)
we now want to use a number of mariadb pods equal to the number of
controllers.

For the scenarios where there is no garbd daemon running we will
now have a "dbmon" OCF script that will ensure that the mariadb
pod on the active controller is part of the primary DB cluster.

(See the "Depends-On" commit messages for more information.)

Depends-On: https://review.openstack.org/650288
Depends-On: https://review.openstack.org/650301

Change-Id: I9a8ef5555a52bd255f81d87cd26a069e43b4d111
Story: 2004712
Task: 30410
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>