Provision dbmon for AIO-DX
For the time being, the dbmon service will be provisioned regardless
stx-openstack application has applied or not. OCF script will handle
the scenario that stx-openstack has not been installed.

A new user story is to be defined to enhance SM to dymamically
provision/deprovision a service. When the enhancement is ready,
the service will be provisioned only after stx-openstack application
is applied.

Story: 2004712
Task: 30363

Depends-On: https://review.openstack.org/#/c/650301/
Depends-On: https://review.openstack.org/#/c/650288/

Change-Id: Ic8ca45c4e9906cb87b92fd8a0fe61ef9134642b6
Signed-off-by: Bin Qian <bin.qian@windriver.com>