Add helm chart for node feature discovery
This packages up the upstream "node feature discovery" project[1] into
a helm chart and includes it in the ISO.  The expectation is that it's
there if end-users want to use it, but it will not be installed by default.

There are two primary files describing Kubernetes resources.  They both
come from version 0.3.0 of the upstream project but one of them needs to
be renamed to work with "helm package" which restricts the allowable filename
suffixes.

[1] https://github.com/kubernetes-sigs/node-feature-discovery/

Change-Id: I6e7c8a629994ad4da3834cbefccb94cd01259cc5
Story: 2005193
Task:  29954