Network agent rebalance/reschedule config changes
In the unlikely event that the network agent
rebalance/reschedule state machines are moving a resource
from one agent to another, and something happens to the host
agent, VIM,  or neutron during this small window of time, it
may be possible for a resource (dhcp network or router) to be
left unattached to an agent.  In order for these resources
to be recovered and reattached to an agent, this update
will turn the network_auto_schedule and router_auto_schedule
neutron config parameters back on.  These were originally
turned off in order to prevent potential race conditions with
the rebalance/reschedule algorithms, but after further
review this was not considered to be an issue. They are
therefore being turned back on.

In addition, during host unlocks the VIM was being restarted
causing the rebalance/reschedule state machines to lose
configuration and potentially uncleanly aborted mid-
operation.  A change is made to only restart the VIM
on initial application apply.

Story: 652151
Task: 30561

Change-Id: I73a069e478200fa4ee9eb959eceecc2dc598eecc
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>