Change the way of disabling ovs container
For the case vswitch_type!='none', ovs doesn't run in contianer. So ovs
pod/container should not run. We controlled ovs container by label, but
a patch[1] broke it.
This patch is to change the method in which we control ovs container.
With this patch, we remove openvswitch chart from compute-kit chart
group so that no ovs container created.

[1]
https://review.openstack.org/#/c/651380/2/kubernetes/applications/stx-openstack/stx-openstack-helm/stx-openstack-helm/manifests/manifest.yaml

Change-Id: I3ba8a3ab45a6e6c1a67b78d335656ed5c0d654a7
Closes-bug: #1824829
Signed-off-by: chengli3 <cheng1.li@intel.com>